requirements.txt
streamlit>=1.35.0
google-genai>=1.0.0
Pillow>=10.0.0
supabase>=2.0.0
